This is a rather large commit.
First, I had to restructure the project by adding src and test folders. I moved all  projects into the src folder. This was necessary for Docker Compose.  It was giving an error about the fact that the Docker Compose project needed to be at a higher level.

launchSettings.json -  a new global launch profile was added for docker-compose. Run the "Docker Compose" launch profile to run the app and sql server containers.

dockerfile - this project file belongs to the web app, and it was completely rewritten to better handle relative pathing.

OdeToFood.App.csproj - I renamed the main app, "OdeToFood.App" to differentiate it from the solution folder itself.

appsettings.json - added new SqlServiceOnDockerConnection" - a database connection string, which was subsequently moved out of here because of the db password.

Program.cs - cleaned up usings, added ApplyDbMigrations method to lay down database and any pending schema changes upon starting. May not be the best idea for production, but it's good to get the db up-to-date after any schema changes since we use code-first approach.

Startup.cs - cleaned up usings, and revised the connection string to be the new SqlServiceOnDockerConnection, along with retry options.

SqlRestaurantData.cs - an error was found for the StringComparision.InvariantCultureIgnoreCase, so I re-wrote the comparison to just compare using a lower case approach.

docker-compose.yml - created initially after adding Container Orchestration Support. But heavily revised with the help of ChatGPT.  There are two services: one for the web app and one for sql server.  SQL Server is persisted through the use of volume. And clients outside of the containers connect over port 1435 instead of 1433, in case a local SSMS client is using 1433. My Sql Server Object Explorer connects to it with "localhost,1435", sa account.

docker-compose.override.yml - file was auto produced when adding Container Orchestration Support.

OdeToFood.sln - updated for new docker-compose.csproj.

There are a couple of external files that manage DB password not stored in the git repo or docker-compose.yml. I bet you can guess which files those are.  If you need to change db settings, remember these two files.

.gitignore - changed to exclude .env files used by docker-compose.yml, but it looks like the .env isn't used because the environment section DBPassword was removed, and the Docker Compose launch profile still starts the containers successfully.